Update RageSoundReader.cpp
The warning inside the loop makes the log too messy and is too much of a performance hit on very low end hardware.

I'm changing it to remove the mid-loop logging and only return a logged warning on a failure.

I'm also re-doing the formatting so it's more compact, but changes nothing functionally.