Removing bApproximate (part 1)
It can't be removed all at once just by deleting every instance of bApproximate, because it often refers to pointers that go elsewhere. so, this is the first part of a multi-stage removal process.

Additional notes: some c-style casts are updated to static cast here, a calculation which should not happen while mutexed is moved, and a macro is removed so that it doesn't hide any errors behind it.