Some maintenance in GameLoop.cpp
Optimize delay loops - instead of running a modulo on an int, just let a uint_8 wrap.

bLogSkips - store the preference in PREFSMAN->m_bLogSkips in a static boolean, instead of calling CheckGameLoopTimerSkips and THEN checking m_bLogSkips.  It's not going to get changed mid-game, so we might as well do this to avoid a function call in the game loop.