RageSoundReader: if loop to replace while loop
Linux users were reporting some versions of GCC failed here unless <cstdint> was included here to use a `std::uint_8` data type, even though GCC 11 / MSVC / XCode accept it as-is.

For simplicity sake, I changed it to an `unsigned`.

Then I realized it seemed clunky to use a `while` loop here, because `while` often doesn't get optimized well by the compiler. It's also a little easier to see what is going on here at a glance with the for loop in place.