RageSoundMixBuffer improvements
Mostly backporting the fixes from the abandoned vector-based RageSoundMixBuffer patch, while retaining the faster malloc/realloc based method.

Mainly, we make Extend noexcept, make more things const so the compiler can better optimize, and ensure every buffer size is a multiple of 1024 - this prevents the memory leak issue in the original SM5 code.

Realloc is now safe instead of malloc since we are ensuring the buffer is the nearest adequate multiple of 1024, a major bug fix (this is actually why the current implementation forcing the buffer to be a similar multiple is functionally solving the issue).

The constructor is similarly modified to give the buffer a default size so there is no risk of Extend attempting to resize a nullptr.

The Extend function uses uint_fast64_t to optimize to whatever data type is best for that platform. It also directly resizes the buffer, similarly to the original SM5 code, as this is fastest.