Use the RageSoundMixBuffer from ITGm 0.7.1
The one fix that has not been ported back to beta from 0.7.1/0.9.0-SHINE is the changed RageSoundMixBuffer.

This PR brings back the RageSoundMixBuffer from that build. That build can be considered very well tested at this point.

I believe it should be ported back, as I believe this has a large part in ensuring that stability.

https://github.com/sukibaby/itgmania/blob/SUKI%60S-FA%2B/src/RageSoundMixBuffer.cpp
https://github.com/sukibaby/itgmania/blob/SUKI%60S-FA%2B/src/RageSoundMixBuffer.h

Remove no longer needed BUF_SIZE variable