Fix display resolution calculation
Sometimes, you might get a weird set of display resolutions in game - this is due to an incorrect implementation of std::ceil (which seems to have been known to be bad, even). We now use std::round instead of std::ceil to ensure the width is correctly rounded to the nearest integer, whereas std::ceil will always round up. We then make sure we got the number we calculated is even. If it isn't, we will increase its value by a factor of one.