Fixed buffer size for RageSoundMixBuffer.
Pre-define a set of audio mix buffer sizes, and let the user choose. By extension, maintain a constant amount of latency, which is greatly beneficial for sync stability. We have known `RageSoundMixBuffer::Extend` is directly responsible for sync stability. The solution is to introduce a new preference to define a fixed audio buffer size. See the PR for more in-depth commentary on these changes.