(Win32 refresh) LoadingWindow_Win32
Just swapping to some more modern ways of doing things, as this code already uses fairly outdated methodology. We can speed up the loading window a bit with these changes. We use Vector and always process three lines at a time in SetText, and use BitBlt instead of SetPixelV for the surface.