Fix RageSoundReader_Merge::Read
Note:  the indentation of the function is messy in order to minimize the amount of lines shown in the diff.  I will later clean up the indentation/whitespace of the function.

This function is responsible for detecting desync and correcting it. A threshold which is "number of frames away from the expected position" we consider tolerable is set to 16 (this is the value it's always been). This is checked inside a loop. We are checking if we are less or equal to the threshold and continuing the loop if so. Because of this, it allows differences up to the threshold limit to go uncorrected. This allows small desyncs to accumulate over time because they will be considered "good enough" until an event exceeding the threshold occurs. This commit changes the logic flow to immediately correct any desync before continuing the loop.