Refactor RageSoundReader_Chain::Read()
Aims to operate in constant time as much as possible by reducing the number of dynamic loops and calculations performed inside loops. The function is refactored to internally track the next sound offset frame to simplify the logic flow. The buffer has been changed from a raw array to a std::array for added safety, and to move away from use of the ARRAYLEN macro. The EOF check was also moved into an existing if-loop checking if m_apActiveSounds is empty, so it's functionally identical. The rest of the function is mostly unchanged, except for a few variable names.